# docs/03_API仕様/openapi.yaml
openapi: 3.1.0
info:
  title: 人事考課システム API
  description: |
    トキワテック人事考課システムのREST API仕様書。

    ## 認証
    Cloudflare Access (JWT) による認証を使用します。
    全てのAPIリクエストには `Authorization: Bearer <token>` ヘッダーが必要です。

    ## 共通仕様
    - 日時はISO 8601形式 (例: `2026-01-24T10:00:00Z`)
    - IDはUUID v4形式
    - ページネーションはoffset/limit方式
  version: 1.0.0
  contact:
    name: トキワテック開発チーム
    email: dev@tokiwa-tech.com
  license:
    name: Proprietary
    identifier: LicenseRef-proprietary
servers:
  - url: https://evaluation.tokiwa-tech.com/api/v1
    description: 本番環境
  - url: https://evaluation-staging.tokiwa-tech.com/api/v1
    description: ステージング環境
  - url: http://localhost:8787/api/v1
    description: ローカル開発環境
...
paths:
...
  /curriculums:
    get:
      tags: [Curriculum]
      summary: カリキュラム一覧取得
      operationId: listCurriculums
      security:
        - BearerAuth: []
      parameters:
        - name: scenarioId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/CurriculumStatus'
        - name: learningPhase
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: カリキュラム一覧
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurriculumListResponse'

    post:
      tags: [Curriculum]
      summary: カリキュラム作成
      description: HR/管理者のみ
      operationId: createCurriculum
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCurriculumRequest'
      responses:
        '201':
          description: 作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurriculumResponse'
  /curriculums/my-assignments:
    get:
      tags: [Curriculum]
      summary: 自分の割り当てカリキュラム一覧
      operationId: listMyAssignments
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 割り当て一覧
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurriculumAssignmentListResponse'
  /curriculums/{curriculumId}:
    get:
      tags: [Curriculum]
      summary: カリキュラム詳細取得
      operationId: getCurriculum
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CurriculumId'
      responses:
        '200':
          description: カリキュラム詳細
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurriculumDetailResponse'

    put:
      tags: [Curriculum]
      summary: カリキュラム更新
      description: HR/管理者のみ
      operationId: updateCurriculum
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CurriculumId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCurriculumRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurriculumResponse'
  /curriculums/{curriculumId}/review:
    post:
      tags: [Curriculum]
      summary: カリキュラムレビュー
      description: HR/管理者のみ
      operationId: reviewCurriculum
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CurriculumId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [approved]
              properties:
                approved:
                  type: boolean
                comment:
                  type: string
      responses:
        '200':
          description: レビュー完了
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurriculumResponse'
  /curriculums/{curriculumId}/publish:
    post:
      tags: [Curriculum]
      summary: カリキュラム公開
      description: HR/管理者のみ
      operationId: publishCurriculum
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CurriculumId'
      responses:
        '200':
          description: 公開成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurriculumResponse'
  /curriculums/{curriculumId}/assign:
    post:
      tags: [Curriculum]
      summary: カリキュラム割り当て
      description: HR/管理者/マネージャーのみ
      operationId: assignCurriculum
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CurriculumId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignCurriculumRequest'
      responses:
        '201':
          description: 割り当て成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurriculumAssignmentResponse'
  /curriculums/{curriculumId}/progress:
    get:
      tags: [Curriculum]
      summary: 学習進捗取得
      operationId: getCurriculumProgress
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CurriculumId'
      responses:
        '200':
          description: 学習進捗
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurriculumProgressResponse'

    put:
      tags: [Curriculum]
      summary: 学習進捗更新
      operationId: updateCurriculumProgress
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CurriculumId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCurriculumProgressRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /curriculums/{curriculumId}/complete:
    post:
      tags: [Curriculum]
      summary: カリキュラム完了
      operationId: completeCurriculum
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CurriculumId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                totalScore:
                  type: number
      responses:
        '200':
          description: 完了処理成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurriculumCompletionResponse'

  # ============================================
  # 昇格 API (将来実装予定)
  # ============================================
  # 注: 昇格機能は将来バージョンで実装予定

  # ============================================
  # Admin API (6 endpoints)
  # ============================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Cloudflare Access JWT token
  parameters:
...
    CurriculumId:
      name: curriculumId
      in: path
      required: true
      schema:
        type: string
      description: カリキュラムID
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: ページ番号
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: 1ページあたりの件数
...
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        hasNext:
          type: boolean

    # ============================================
    # 列挙型
    # ============================================
...
    CurriculumStatus:
      type: string
      enum: [DRAFT, GENERATED, REVIEWED, PUBLISHED, ARCHIVED]
...
    Curriculum:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        targetGrades:
          type: array
          items:
            type: string
        estimatedHours:
          type: integer
        status:
          type: string
    AssignCurriculumRequest:
      type: object
      required: [curriculumId, employeeIds]
      properties:
        curriculumId:
          type: string
        employeeIds:
          type: array
          items:
            type: string
        deadline:
          type: string
          format: date
...
    CurriculumListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Curriculum'
        pagination:
          $ref: '#/components/schemas/Pagination'
    CurriculumResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Curriculum'
    CurriculumDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          allOf:
            - $ref: '#/components/schemas/Curriculum'
            - type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/CurriculumItem'
    CurriculumItem:
      type: object
      properties:
        id:
          type: string
        itemNumber:
          type: integer
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [VIDEO, DOCUMENT, QUIZ, EXERCISE, ASSIGNMENT]
        duration:
          type: integer
        passScore:
          type: integer
          nullable: true
    CreateCurriculumRequest:
      type: object
      required: [scenarioId, name]
      properties:
        scenarioId:
          type: string
        name:
          type: string
        description:
          type: string
        targetGrades:
          type: array
          items:
            type: string
        learningPhase:
          type: string
        estimatedHours:
          type: integer
        noiseTypes:
          type: array
          items:
            type: string
        llmModel:
          type: string
    UpdateCurriculumRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        targetGrades:
          type: array
          items:
            type: string
        learningPhase:
          type: string
        estimatedHours:
          type: integer
        status:
          $ref: '#/components/schemas/CurriculumStatus'
    CurriculumAssignmentListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/CurriculumAssignment'
        pagination:
          $ref: '#/components/schemas/Pagination'
    CurriculumAssignment:
      type: object
      properties:
        id:
          type: string
        curriculumId:
          type: string
        employeeId:
          type: string
        assignedBy:
          type: string
        assignedAt:
          type: string
          format: date-time
        deadline:
          type: string
          format: date
          nullable: true
        passThreshold:
          type: integer
        progress:
          type: integer
        status:
          type: string
          enum: [ASSIGNED, IN_PROGRESS, COMPLETED, OVERDUE]
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        totalScore:
          type: number
          nullable: true
        isPassed:
          type: boolean
          nullable: true
        progressItems:
          type: array
          items:
            type: object
    CurriculumAssignmentResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/CurriculumAssignment'
    CurriculumProgressResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            curriculumId:
              type: string
            assignmentId:
              type: string
            progress:
              type: integer
            status:
              type: string
            progressItems:
              type: array
              items:
                type: object
    UpdateCurriculumProgressRequest:
      type: object
      properties:
        curriculumItemId:
          type: string
        status:
          type: string
          enum: [NOT_STARTED, IN_PROGRESS, COMPLETED]
        score:
          type: number
        maxScore:
          type: number
        attempts:
          type: integer
        timeSpentMinutes:
          type: integer
        notes:
          type: string
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        progress:
          type: integer
    CurriculumCompletionResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            curriculumId:
              type: string
            assignmentId:
              type: string
            status:
              type: string
            totalScore:
              type: number
            isPassed:
              type: boolean